Move ForgotPasswordForm component to app folder
